+++ src/dudu/cli_usage.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace dudu {
+
+void print_cli_usage(bool project_driver);
+void print_cli_version();
+
+} // namespace dudu